chore: run tests in ci (#53)
This PR ensures that all tests in `Tests.lean` cause an error when there
is a test failure by using `#guard_msgs`, that `Tests.lean` and
`Example.lean` are run by `lake test` and that CI runs `lake test`. This
should hopefully prevent situations like the one in #51 from occurring
in the future.

It also replaces `Lean.RBTree` with `Std.TreeSet`.